+++ phono3py/api_phono3py.py
@@ -59,7 +59,7 @@ from phonopy.interface.mlp import PhonopyMLP
 from phonopy.interface.pypolymlp import (
     PypolymlpParams,
 )
-from phonopy.interface.symfc import SymfcFCSolver
+from phonopy.interface.symfc import SymfcFCSolver, symmetrize_by_projector
 from phonopy.physical_units import get_physical_units
 from phonopy.structure.atoms import PhonopyAtoms
 from phonopy.structure.cells import (
@@ -744,16 +744,21 @@ class Phono3py:
         self._phonon_mlp_dataset = mlp_dataset
 
     @property
-    def mlp(self) -> Optional[PhonopyMLP]:
+    def mlp(self) -> PhonopyMLP | None:
         """Setter and getter of PhonopyMLP dataclass."""
         return self._mlp
 
     @mlp.setter
-    def mlp(self, mlp) -> Optional[PhonopyMLP]:
+    def mlp(self, mlp: PhonopyMLP):
         self._mlp = mlp
 
     @property
-    def band_indices(self) -> list[NDArray]:
+    def phonon_mlp(self):
+        """Return MLP instance for fc2."""
+        return self._phonon_mlp
+
+    @property
+    def band_indices(self) -> list[NDArray] | None:
         """Setter and getter of band indices.
 
         list[NDArray]
@@ -811,7 +816,7 @@ class Phono3py:
         return self._supercells_with_displacements
 
     @property
-    def phonon_supercells_with_displacements(self):
+    def phonon_supercells_with_displacements(self) -> list[PhonopyAtoms] | None:
         """Return supercells with displacements for fc2.
 
         list of PhonopyAtoms
@@ -1084,16 +1089,6 @@ class Phono3py:
         """
         return self._bz_grid
 
-    @property
-    def mlp(self):
-        """Return MLP instance."""
-        return self._mlp
-
-    @property
-    def phonon_mlp(self):
-        """Return MLP instance for fc2."""
-        return self._phonon_mlp
-
     def init_phph_interaction(
         self,
         nac_q_direction=None,
@@ -1253,13 +1248,14 @@ class Phono3py:
 
     def generate_displacements(
         self,
-        distance: Optional[float] = None,
-        cutoff_pair_distance: Optional[float] = None,
-        is_plusminus: Union[bool, str] = "auto",
+        distance: float | None = None,
+        cutoff_pair_distance: float | None = None,
+        is_plusminus: bool | str = "auto",
         is_diagonal: bool = True,
-        number_of_snapshots: Optional[Union[int, Literal["auto"]]] = None,
-        random_seed: Optional[int] = None,
-        max_distance: Optional[float] = None,
+        number_of_snapshots: int | Literal["auto"] | None = None,
+        random_seed: int | None = None,
+        max_distance: float | None = None,
+        number_estimation_factor: float | None = None,
     ):
         """Generate displacement dataset in supercell for fc3.
 
@@ -1307,7 +1303,7 @@ class Phono3py:
             pair of displacements is considered to calculate fc3 or not. Default
             is None, which means cutoff is not used.
         is_plusminus : True, False, or 'auto', optional
-            With True, atomis are displaced in both positive and negative
+            With True, atoms are displaced in both positive and negative
             directions. With False, only one direction. With 'auto', mostly
             equivalent to is_plusminus=True, but only one direction is chosen
             when the displacements in both directions are symmetrically
@@ -1328,9 +1324,15 @@ class Phono3py:
         random_seed : int or None, optional
             Random seed for random displacements generation. Default is None.
         max_distance : float or None, optional
-            In random direction and distance displacements generation, this
-            value is specified. In random direction and random distance
-            displacements generation, this value is used as `max_distance`.
+            When specified, displacements are generated with random direction
+            and random distance. This value serves as the maximum distance,
+            while the `distance` parameter sets the minimum distance. The
+            displacement distance is randomly sampled from a uniform
+            distribution between these two bounds.
+        number_estimation_factor : float, optional
+            This factor multiplies the number of snapshots estimated by symfc
+            when `number_of_snapshots` is set to "auto". Default is None, which
+            sets this factor to 8 when `max_distance` is specified, otherwise 4.
 
         """
         if distance is None:
@@ -1346,14 +1348,19 @@ class Phono3py:
                     options = None
                 else:
                     options = {"cutoff": {3: cutoff_pair_distance}}
-                _number_of_snapshots = (
-                    SymfcFCSolver(
-                        self._supercell,
-                        symmetry=self._symmetry,
-                        options=options,
-                    ).estimate_numbers_of_supercells(orders=[3])[3]
-                    * 2
-                )
+                _number_of_snapshots = SymfcFCSolver(
+                    self._supercell,
+                    symmetry=self._symmetry,
+                    options=options,
+                ).estimate_numbers_of_supercells(orders=[3])[3]
+                if number_estimation_factor is None:
+                    if max_distance is None:
+                        _number_of_snapshots *= 4
+                    else:
+                        _number_of_snapshots *= 8
+                else:
+                    _number_of_snapshots *= number_estimation_factor
+                    _number_of_snapshots = int(_number_of_snapshots)
             else:
                 _number_of_snapshots = number_of_snapshots
             self._dataset = self._generate_random_displacements(
@@ -1484,8 +1491,9 @@ class Phono3py:
         self,
         symmetrize_fc3r: bool = False,
         is_compact_fc: bool = False,
-        fc_calculator: Optional[Union[str, dict]] = None,
-        fc_calculator_options: Optional[Union[str, dict]] = None,
+        fc_calculator: Literal["traditional", "symfc", "alm"] | None = None,
+        fc_calculator_options: str | None = None,
+        use_symfc_projector: bool = False,
     ):
         """Calculate fc3 from displacements and forces.
 
@@ -1506,8 +1514,11 @@ class Phono3py:
             cells. Default is False.
         fc_calculator : str, optional
             Force constants calculator given by str.
-        fc_calculator_options : dict or str, optional
+        fc_calculator_options : str, optional
             Options for external force constants calculator.
+        use_symfc_projector : bool, optional
+            If True, the force constants are symmetrized by symfc projector
+            instead of traditional approach.
 
         """
         fc_solver_name = fc_calculator if fc_calculator is not None else "traditional"
@@ -1525,17 +1536,44 @@ class Phono3py:
         fc2 = fc_solver.force_constants[2]
         fc3 = fc_solver.force_constants[3]
 
-        if fc_calculator is None or fc_calculator == "traditional":
-            if symmetrize_fc3r:
+        if symmetrize_fc3r:
+            if use_symfc_projector and fc_calculator is None:
+                if self._log_level:
+                    print("Symmetrizing fc3 by symfc projector.", flush=True)
+                fc3 = symmetrize_by_projector(
+                    self._supercell,
+                    fc3,
+                    3,
+                    primitive=self._primitive,
+                    log_level=self._log_level,
+                    show_credit=True,
+                )
+                if self._fc2 is None:
+                    if self._log_level:
+                        print("Symmetrizing fc2 by symfc projector.", flush=True)
+                    fc2 = symmetrize_by_projector(
+                        self._supercell,
+                        fc2,
+                        2,
+                        primitive=self._primitive,
+                        log_level=self._log_level,
+                    )
+            elif fc_calculator is None or fc_calculator == "traditional":
+                if self._log_level:
+                    print("Symmetrizing fc3 by traditional approach.", flush=True)
                 if is_compact_fc:
                     set_translational_invariance_compact_fc3(fc3, self._primitive)
                     set_permutation_symmetry_compact_fc3(fc3, self._primitive)
-                    if self._fc2 is None:
-                        symmetrize_compact_force_constants(fc2, self._primitive)
                 else:
                     set_translational_invariance_fc3(fc3)
                     set_permutation_symmetry_fc3(fc3)
-                    if self._fc2 is None:
+
+                if self._fc2 is None:
+                    if self._log_level:
+                        print("Symmetrizing fc2 by traditional approach.", flush=True)
+                    if is_compact_fc:
+                        symmetrize_compact_force_constants(fc2, self._primitive)
+                    else:
                         symmetrize_force_constants(fc2)
 
         self._fc3 = fc3
@@ -1573,10 +1611,11 @@ class Phono3py:
 
     def produce_fc2(
         self,
-        symmetrize_fc2=False,
-        is_compact_fc=False,
-        fc_calculator=None,
-        fc_calculator_options=None,
+        symmetrize_fc2: bool = False,
+        is_compact_fc: bool = False,
+        fc_calculator: Literal["traditional", "symfc", "alm"] | None = None,
+        fc_calculator_options: str | None = None,
+        use_symfc_projector: bool = False,
     ):
         """Calculate fc2 from displacements and forces.
 
@@ -1597,8 +1636,11 @@ class Phono3py:
             cells. Default is False.
         fc_calculator : str or None
             Force constants calculator given by str.
-        fc_calculator_options : dict
+        fc_calculator_options : str or None
             Options for external force constants calculator.
+        use_symfc_projector : bool, optional
+            If True, the force constants are symmetrized by symfc projector
+            instead of traditional approach.
 
         """
         if self._phonon_dataset is None:
@@ -1606,6 +1648,8 @@ class Phono3py:
         else:
             disp_dataset = self._phonon_dataset
 
+        if disp_dataset is None:
+            raise RuntimeError("Displacement dataset is not set.")
         if not forces_in_dataset(disp_dataset):
             raise RuntimeError("Forces are not set in the dataset.")
 
@@ -1620,8 +1664,16 @@ class Phono3py:
             log_level=self._log_level,
         )
 
-        if fc_calculator is None or fc_calculator == "traditional":
-            if symmetrize_fc2:
+        if symmetrize_fc2 and (fc_calculator is None or fc_calculator == "traditional"):
+            if use_symfc_projector:
+                self._fc2 = symmetrize_by_projector(
+                    self._phonon_supercell,
+                    self._fc2,
+                    2,
+                    primitive=self._primitive,
+                    log_level=self._log_level,
+                )
+            else:
                 if is_compact_fc:
                     symmetrize_compact_force_constants(
                         self._fc2, self._phonon_primitive
@@ -2229,7 +2281,7 @@ class Phono3py:
             )
 
     def save(
-        self, filename: str = "phono3py_params.yaml", settings: Optional[dict] = None
+        self, filename: str = "phono3py_params.yaml", settings: dict | None = None
     ):
         """Save parameters in Phono3py instants into file.
 
@@ -2256,7 +2308,7 @@ class Phono3py:
 
     def develop_mlp(
         self,
-        params: Optional[Union[PypolymlpParams, dict, str]] = None,
+        params: PypolymlpParams | dict | str | None = None,
         test_size: float = 0.1,
     ):
         """Develop machine learning potential.
@@ -2473,7 +2525,7 @@ class Phono3py:
 
     def _build_phonon_supercells_with_displacements(
         self, supercell: PhonopyAtoms, dataset
-    ):
+    ) -> list[PhonopyAtoms]:
         supercells = []
         positions = supercell.positions
         magmoms = supercell.magnetic_moments
@@ -2733,8 +2785,8 @@ class Phono3py:
         number_of_atoms: int,
         distance: float = 0.03,
         is_plusminus: bool = False,
-        random_seed: Optional[int] = None,
-        max_distance: Optional[float] = None,
+        random_seed: int | None = None,
+        max_distance: float | None = None,
     ):
         if random_seed is not None and random_seed >= 0 and random_seed < 2**32:
             _random_seed = random_seed