+++ phono3py/scripts/phono3py.py
@@ -37,9 +37,5 @@ from phono3py.cui.phono3py_script import main
 
 def run():
     """Run phono3py script."""
-    argparse_control = {
-        "fc_symmetry": False,
-        "is_nac": False,
-        "load_phono3py_yaml": False,
-    }
+    argparse_control = {"load_phono3py_yaml": False}
     main(**argparse_control)